Splits sqs.Driver into multiple files
